Remove manual flags for dynamic linking of the Windows CRT library
The use of `/MDd` for the Debug compilation type and `/MD` for all the other is already ensured by the `Windows-MSVC` CMake Platform file,
so this line should not be necessary.

Furthermore, on Multi Config CMake Generators such as Visual Studio, the `CMAKE_BUILD_TYPE` variables does not contain the actual configuration with which the code is compiled, that is instead specified by the user at compile time rather than configuration time. This mean that
the `/MD` flag (linking with the Non-Debug CRT Dynamic Library) is passed even to Debug builds, and this causes linking error with third-party debug dependencies in gazebo .

[1] https://github.com/Kitware/CMake/blob/6dc7262bf77f551a7f3c36d7ca3023c63ef20af8/Modules/Platform/Windows-MSVC.cmake#L292
[2] https://cmake.org/Wiki/CMake_FAQ#How_can_I_build_my_MSVC_application_with_a_static_runtime.3F