Fix compiler warning and signedness issue
This commit fixes a compiler warning I was getting with regard to signed
and unsigned types. On line 794 of `Discovery.hh` we use the recvfrom
but we use an unsigned int to check if the error is correct or not. By
default recvfrom returns an `ssize_t` not a `uint16_t`. It seems a fix
for this has landed in garden but has not been backported to citadel or
fortress.

Signed-off-by: Arjo Chakravarty <arjoc@intrinsic.ai>