Keep signature of `Node::Shared` to maintain ABI stability
`Node::Shared` is a private function but it's used in `details/Node.hh`
in templates which means there could be calls to `Node::Shared` embedded
in other libraries. Since we're mainly tracking the number of `Node`
instances, and since `Node::Shared` is private, it's okay to use raw pointers
internal use.

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>