Use `std::shared_ptr` for `gz::transport::NodeShared` (#484)
Currently, `NodeShared` is never destroyed once it's created. This makes
it hard for writing tests that do not interfere with each other. This
patch uses a reference counted smart pointer (`std::shared_ptr`) to keep
track of `NodeShared` instances, so that it gets properly destroyed when
when all `Node` instances, which themselves contain `NodeShared`
instances are destroyed.

---------

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>