Add an explicit dependency on cppzmq-dev
In Jammy, the zmq C++ headers were included in libzmq-dev. However, in
Noble, the headers are now in a separate package, `cppzmq-dev`,
and that package is recommended by libzmq-dev. This shouldn't affect
most users, but anywhere these dependencies are installed with
`--no-install-recommends` will now break because it won't have
cppzmq-dev installed. This patch just adds an explicit dependency to
avoid the problem.

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>